Revert "refactor: remove missed redundant computation of `projectRoot` (#7421)"
This reverts commit bea65583f72fcf8346ed82d4b9b1176669bcfd23.
