+++ fixtures/workflow/package.json
@@ -4,7 +4,7 @@
 	"scripts": {
 		"deploy": "wrangler deploy",
 		"start": "wrangler dev",
-		"test:ci": "vitest run"
+		"test:ci": "vitest"
 	},
 	"devDependencies": {
 		"@cloudflare/workers-types": "^4.20241106.0",
