+++ packages/wrangler/src/__tests__/get-entry.test.ts
@@ -1,15 +1,18 @@
 import path from "path";
 import dedent from "ts-dedent";
-import { normalizeAndValidateConfig } from "../config/validation";
+import { defaultWranglerConfig } from "../config/config";
 import { getEntry } from "../deployment-bundle/entry";
 import { mockConsoleMethods } from "./helpers/mock-console";
 import { runInTempDir } from "./helpers/run-in-tmp";
 import { seed } from "./helpers/seed";
-import type { Config, RawConfig } from "../config";
+import type { Config } from "../config/config";
 import type { Entry } from "../deployment-bundle/entry";
 
-function getConfig(raw: RawConfig = {}, configPath?: string): Config {
-	return normalizeAndValidateConfig(raw, configPath, {}).config;
+function getConfig(): Config {
+	return {
+		...defaultWranglerConfig,
+		projectRoot: process.cwd(),
+	};
 }
 
 function normalize(entry: Entry): Entry {
@@ -80,7 +83,11 @@ describe("getEntry()", () => {
 							}
 						`,
 		});
-		const entry = await getEntry({}, getConfig({ main: "index.ts" }), "deploy");
+		const entry = await getEntry(
+			{},
+			{ ...getConfig(), main: "index.ts" },
+			"deploy"
+		);
 		expect(normalize(entry)).toMatchObject({
 			file: "/tmp/dir/index.ts",
 			moduleRoot: "/tmp/dir",
@@ -99,7 +106,7 @@ describe("getEntry()", () => {
 		});
 		const entry = await getEntry(
 			{},
-			getConfig({ main: "src/index.ts" }),
+			{ ...getConfig(), main: "src/index.ts" },
 			"deploy"
 		);
 		expect(normalize(entry)).toMatchObject({
@@ -120,12 +127,12 @@ describe("getEntry()", () => {
 		});
 		const entry = await getEntry(
 			{},
-			getConfig(
-				{
-					main: "src/index.ts",
-				},
-				"other-worker/wrangler.toml"
-			),
+			{
+				...getConfig(),
+				main: "src/index.ts",
+				configPath: "other-worker/wrangler.toml",
+				projectRoot: "other-worker",
+			},
 			"deploy"
 		);
 		expect(normalize(entry)).toMatchObject({
